requirements.txt
openai-whisper==20250625
torch>=2.0.0
numpy>=1.20.0
tqdm>=4.60.0
more-itertools>=8.0.0
tiktoken>=0.3.0
numba>=0.56.0
